# scan-engine/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "scan-to-map-studio API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/groups/{name}": {
      "get": {
        "summary": "Api Group",
        "operationId": "api_group_api_groups__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Api Delete Group",
        "operationId": "api_delete_group_api_groups__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/groups/{name}": {
      "get": {
        "summary": "Group Workspace",
        "operationId": "group_workspace_groups__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects": {
      "get": {
        "summary": "Api List Projects",
        "operationId": "api_list_projects_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectEntry"
                  },
                  "type": "array",
                  "title": "Response Api List Projects Api Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Api Create Project",
        "operationId": "api_create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_create_project_api_projects_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{name}/process": {
      "post": {
        "summary": "Api Process Project",
        "operationId": "api_process_project_api_projects__name__process_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_process_project_api_projects__name__process_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessStarted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{name}/status": {
      "get": {
        "summary": "Api Get Status",
        "operationId": "api_get_status_api_projects__name__status_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{name}": {
      "delete": {
        "summary": "Api Delete Project",
        "operationId": "api_delete_project_api_projects__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{name}/align/geojson": {
      "post": {
        "summary": "Api Align Geojson",
        "operationId": "api_align_geojson_api_projects__name__align_geojson_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_align_geojson_api_projects__name__align_geojson_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api Align Geojson Api Projects  Name  Align Geojson Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{name}/align/image": {
      "post": {
        "summary": "Api Align Image",
        "operationId": "api_align_image_api_projects__name__align_image_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_align_image_api_projects__name__align_image_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api Align Image Api Projects  Name  Align Image Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_api_align_geojson_api_projects__name__align_geojson_post": {
        "properties": {
          "geojson": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Geojson"
          }
        },
        "type": "object",
        "required": [
          "geojson"
        ],
        "title": "Body_api_align_geojson_api_projects__name__align_geojson_post"
      },
      "Body_api_align_image_api_projects__name__align_image_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_api_align_image_api_projects__name__align_image_post"
      },
      "Body_api_create_project_api_projects_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Body_api_create_project_api_projects_post"
      },
      "Body_api_process_project_api_projects__name__process_post": {
        "properties": {
          "usdz": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usdz"
          },
          "scan_file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan File"
          },
          "robot_map_pgm": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Robot Map Pgm"
          },
          "robot_map_yaml": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Robot Map Yaml"
          },
          "trajectory": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trajectory"
          },
          "remove_isolated_clusters": {
            "type": "boolean",
            "title": "Remove Isolated Clusters",
            "default": false
          },
          "isolated_cluster_min_area": {
            "type": "number",
            "title": "Isolated Cluster Min Area",
            "default": 0.3
          },
          "classify": {
            "type": "boolean",
            "title": "Classify",
            "default": false
          }
        },
        "type": "object",
        "title": "Body_api_process_project_api_projects__name__process_post"
      },
...
      "GroupStatus": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dir": {
            "type": "string",
            "title": "Dir"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "scans": {
            "items": {
              "$ref": "#/components/schemas/ScanStatus"
            },
            "type": "array",
            "title": "Scans"
          },
          "has_alignment": {
            "type": "boolean",
            "title": "Has Alignment"
          },
          "has_merged": {
            "type": "boolean",
            "title": "Has Merged"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "has_floor": {
            "type": "boolean",
            "title": "Has Floor",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "dir",
          "reference",
          "scans",
          "has_alignment",
          "has_merged",
          "ready"
        ],
        "title": "GroupStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessStarted": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "type": {
            "type": "string",
            "enum": [
              "single",
              "group"
            ],
            "title": "Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "has_floorplan": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Floorplan"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "group_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Url"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "status",
          "type"
        ],
        "title": "ProcessStarted",
        "description": "POST /api/projects/{name}/process: a single scan starts the pipeline; a multi-scan zip becomes a group."
      },
      "ProjectCreated": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreated"
      },
      "ProjectEntry": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "steps": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Steps",
            "default": {}
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "mtime": {
            "type": "number",
            "title": "Mtime"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "name",
          "phase",
          "mtime"
        ],
        "title": "ProjectEntry",
        "description": "GET /api/projects item (== studio.project.list_projects)."
      },
      "ProjectStatus": {
        "properties": {
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "steps": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "pending",
                "active",
                "done",
                "skip",
                "error"
              ]
            },
            "type": "object",
            "title": "Steps",
            "default": {}
          },
          "log": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Log",
            "default": []
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ProjectStatus",
        "description": "status.json (== studio.status.read_status); phase None = never processed."
      },
...
      "ScanStatus": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "has_folder": {
            "type": "boolean",
            "title": "Has Folder"
          },
          "has_usdz": {
            "type": "boolean",
            "title": "Has Usdz"
          },
          "has_project": {
            "type": "boolean",
            "title": "Has Project"
          },
          "has_slice": {
            "type": "boolean",
            "title": "Has Slice"
          },
          "method": {
            "type": "string",
            "title": "Method"
          }
        },
        "type": "object",
        "required": [
          "id",
          "has_folder",
          "has_usdz",
          "has_project",
          "has_slice",
          "method"
        ],
        "title": "ScanStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}